Fixed some bugs found in GEN_FILE
genfile commands would still not terminate properly
or would end up de-synced with the host in one way
or another.

the GEN_FILE sub-protocol has been updated to
address these issues by having the direction of the
genfile data defined in NEW_CMD so this client will
now know if the command uploads to the host or
downloads from the host when it is defined instead
of trying to determine that at runtime.

updated all documentation related to the GEN_FILE
sub-protocol to reflect these changes.